% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BileAcid.R
\docType{data}
\name{BileAcid}
\alias{BileAcid}
\title{Bile Acid Data}
\format{
A data frame with 198 rows and 34 variables. Each row is a patient id
and each column is an bile acid measurement.
}
\usage{
BileAcid
}
\description{
A bile acid data set taken from Lei et al. (2018)
\doi{https://doi.org/10.1096/fj.201700055R} via Wei et al. (2018)
\doi{https://doi.org/10.1371/journal.pcbi.1005973}
(corresponding GitHub repository: https://github.com/WandeRum/GSimp). The 
values in \code{BileAcid} can be assumed to be independent log-normal 
measurements.
}
\keyword{datasets}
