% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittingMethods.R
\name{ConvexPrimal}
\alias{ConvexPrimal}
\title{Convex Optimization of the Kiefer-Wolfowitz NPMLE}
\usage{
ConvexPrimal(A, ...)
}
\arguments{
\item{A}{numeric matrix likelihoods}

\item{...}{further arguments passed to \code{Rmosek} such as \code{rtol}}
}
\value{
a vector containing the fitted prior
}
\description{
This method only works if there is a working installation of \code{REBayes}
available. See the \code{REBayes} package and corresponding papers for more
implementation details.
}
\details{
The matrix \code{A} is structured as follows: A_ij = P(X_i | theta = t_j),
where X_i is the i'th observation and t_j is the j'th set of
parameters/grid-point.
}
