% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EM}
\alias{EM}
\title{Nonparametric Maximum Likelihood via Expectation Maximization}
\usage{
EM(A, maxiter = 1000000L, tol = 1e-06)
}
\arguments{
\item{A}{numeric matrix likelihoods}

\item{maxiter}{early stopping condition}

\item{tol}{convergence tolerance}
}
\value{
the estimated prior distribution (a vector of masses corresponding
to the columns of A)
}
\description{
Compute the nonparametric maximum likelihood estimate given a likelihood
matrix. The matrix A is structured so that A_\{ij\} = f(X_i | theta_j) for
some grid of potential parameter values theta_1, ..., theta_p and
observations X_1, ..., X_n. The parameters, theta_j, can be multidimensional
because all that is required is the likelihood. Convergence is achieved when
the relative improvements of the log-likelihood is below the provided
tolerance level.
}
\examples{
set.seed(1)
t = sample.int(5, size = 1000, replace = TRUE)
x = t + stats::rnorm(1000)
gr = seq(from = min(x), to = max(x), length.out = 500)
A = stats::dnorm(outer(x, gr, "-"))
EM(A)

\dontrun{
all.equal(
    REBayes::KWPrimal(A = A, d = rep(1, 500), w = rep(1/1000, 1000))$f,
    drop(EM(A, maxiter = 1e+9, tol = 1e-16)) # EM alg converges slowly
)
}

}
