% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{fitted.EBGaME}
\alias{fitted.EBGaME}
\title{Fitted Estimates of an EBGaME object}
\usage{
\method{fitted}{EBGaME}(object, method = "mean", ...)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{EBGaME}}}

\item{method}{either "mean" or "L1mediod" corresponding to the S3 methods:
\code{posterior_*.EBGaME()}}

\item{...}{not used}
}
\description{
Compute either the posterior mean (default) or posterior L1 mediod which
corresponds to the posterior median in one-dimension.
}
