% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{fitted.ebTobit}
\alias{fitted.ebTobit}
\title{Fitted Estimates of an ebTobit object}
\usage{
\method{fitted}{ebTobit}(object, method = "mean", ...)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{ebTobit}}}

\item{method}{either "mean", "L1mediod", or "mode" corresponding to the 
methods: \code{posterior_*.ebTobit()}}

\item{...}{not used}
}
\value{
matrix containing the posterior estimates for measurements in the
fit empirical Bayes model \code{object}
}
\description{
Compute either the posterior mean (default) or posterior L1 mediod which
corresponds to the posterior median in one-dimension.
}
