% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{is.ebTobit}
\alias{is.ebTobit}
\title{Validate ebTobit Object}
\usage{
is.ebTobit(object)
}
\arguments{
\item{object}{any R object}
}
\value{
boolean: TRUE if the object is a valid \code{\link{ebTobit}} object
}
\description{
Validate ebTobit Object
}
