% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{logLik.EBayesMat}
\alias{logLik.EBayesMat}
\title{Marginal Log-likelihood of an EBayesMat object}
\usage{
\method{logLik}{EBayesMat}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{EBayesMat}}}

\item{...}{not used}
}
\description{
Marginal Log-likelihood of an EBayesMat object
}
