% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{logLik.EBGaME}
\alias{logLik.EBGaME}
\title{Marginal Log-likelihood of an EBGaME object}
\usage{
\method{logLik}{EBGaME}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{EBGaME}}}

\item{...}{not used}
}
\description{
Marginal Log-likelihood of an EBGaME object
}
