% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBGaME.R
\name{new_EBGaME}
\alias{new_EBGaME}
\title{Create a new EBGaME object}
\usage{
new_EBGaME(prior, gr, lik)
}
\arguments{
\item{prior}{numeric vector of non-negative weights (sums to one)}

\item{gr}{numeric matrix of support points}

\item{lik}{numeric matrix of likelihoods}
}
\value{
an \code{EBGaME} object containing at least the prior weights,
corresponding grid/support points, and likelihood matrix relating the grid to
the observations
}
\description{
Validate the provided elements and populate the object. Future extensions may
be loosened to allow non-numeric \code{gr} but it is currently required for
posterior mean and mediod calculations.
}
