% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebTobit.R
\name{new_ebTobit}
\alias{new_ebTobit}
\title{Create a new ebTobit object}
\usage{
new_ebTobit(prior, gr, lik)
}
\arguments{
\item{prior}{numeric vector of non-negative weights (sums to one)}

\item{gr}{numeric matrix of support points}

\item{lik}{numeric matrix of likelihoods}
}
\value{
an \code{EBayesMat} object containing at least the prior weights,
corresponding grid/support points, and likelihood matrix relating the grid to
the observations
}
\description{
Validate the provided elements and populate the object. Current methods
require that \code{gr} is numeric for that calculation of posterior
statistics (mean and mediod).
}
