% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{posterior_mean.EBayesMat}
\alias{posterior_mean.EBayesMat}
\title{Compute Posterior Mean of an EBayesMat object}
\usage{
posterior_mean.EBayesMat(object)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{EBayesMat}}}
}
\description{
Compute Posterior Mean of an EBayesMat object
}
