% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{posterior_mean.EBGaME}
\alias{posterior_mean.EBGaME}
\title{Compute Posterior Mean of an EBGaME object}
\usage{
posterior_mean.EBGaME(object)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{EBGaME}}}
}
\description{
Compute Posterior Mean of an EBGaME object
}
