% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{posterior_mode.EBayesMat}
\alias{posterior_mode.EBayesMat}
\title{Compute Posterior Mode of an EBayesMat object}
\usage{
posterior_mode.EBayesMat(object)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{EBayesMat}}}
}
\description{
Compute Posterior Mode of an EBayesMat object
}
