% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{predict.EBayesMat}
\alias{predict.EBayesMat}
\title{Fitted Estimates of an EBayesMat object}
\usage{
\method{predict}{EBayesMat}(object, L, R = L, method = "mean", ...)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{EBayesMat}}}

\item{L}{n x p matrix of lower bounds on observations}

\item{R}{n x p matrix of upper bounds on observations}

\item{method}{either "mean", "L1mediod", or "mode" corresponding to the 
methods: \code{posterior_*.EBayesMat()}}

\item{...}{not used}
}
\description{
Compute either the posterior mean (default) or posterior L1 mediod which
corresponds to the posterior median in one-dimension.
}
