% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{predict.ebTobit}
\alias{predict.ebTobit}
\title{Fitted Estimates of an ebTobit object}
\usage{
\method{predict}{ebTobit}(object, L, R = L, s1 = 1, method = "mean", ...)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{ebTobit}}}

\item{L}{n x p matrix of lower bounds on observations}

\item{R}{n x p matrix of upper bounds on observations}

\item{s1}{a single numeric standard deviation or an n x p matrix of standard
deviations}

\item{method}{either "mean", "L1mediod", or "mode" corresponding to the 
methods: \code{posterior_*.ebTobit()}}

\item{...}{not used}
}
\value{
matrix of posterior estimates for new observations under the
provided, pre-fit empirical Bayes model \code{object}
}
\description{
Compute either the posterior mean (default) or posterior L1 mediod which
corresponds to the posterior median in one-dimension.
}
