% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{logLik.ebTobit}
\alias{logLik.ebTobit}
\title{Marginal Log-likelihood of an ebTobit object}
\usage{
\method{logLik}{ebTobit}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{ebTobit}}}

\item{...}{not used}
}
\value{
log likelihood for the fitted empirical Bayes model in \code{object}
}
\description{
Marginal Log-likelihood of an ebTobit object
}
