% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{posterior_mean.ebTobit}
\alias{posterior_mean.ebTobit}
\title{Compute Posterior Mean of an ebTobit object}
\usage{
posterior_mean.ebTobit(object)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{ebTobit}}}
}
\value{
numeric matrix of posterior means for the fitted empirical Bayes
model in \code{object}
}
\description{
Compute Posterior Mean of an ebTobit object
}
