% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3utils.R
\name{posterior_mode.ebTobit}
\alias{posterior_mode.ebTobit}
\title{Compute Posterior Mode of an ebTobit object}
\usage{
posterior_mode.ebTobit(object)
}
\arguments{
\item{object}{an object inheriting from class \code{\link{ebTobit}}}
}
\value{
numeric matrix of posterior modes for the fitted empirical
Bayes model in \code{object}
}
\description{
Compute Posterior Mode of an ebTobit object
}
